backend/pip_requirements.txt
alembic==1.6.5
botocore==1.20.27
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cognitojwt==1.3.0
ecdsa==0.14.1
Flask==2.0.0
Flask-Cognito==1.17
Flask-Cors==3.0.10
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==2.0.0
pyasn1==0.4.8
pynamodb==5.0.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.17
urllib3==1.26.4
Werkzeug==2.0.0
